refactor(config): split monolithic schema into per-feature modules
